backend/pyproject.toml
[tool.poetry]
name = "goal-planner"
version = "0.1.0"
description = ""
authors = ["shurxet <s.shurupov.88@mail.ru>"]
readme = "README.md"
packages = [{include = "goal_planner"}]

[tool.poetry.dependencies]
python = "^3.11.3"
django = "4.2"
django-environ = "^0.10.0"
phonenumbers = "8.13.8"
psycopg2-binary = "^2.9.5"
django-phonenumber-field = "^7.0.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
drf-spectacular = {extras = ["sidecar"], version = "^0.26.1"}
django-cors-headers = "3.14.0"
drf-spectacular-sidecar = "2023.4.1"
social-auth-app-django = "^5.2.0"
django-filter = "^23.1"
pydantic = "^1.10.7"
requests = "^2.28.2"
telebot = "^0.0.5"


[tool.poetry.group.dev.dependencies]
ansible-vault-win = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
